move data models out of the single file
