+++ SongPlayHistory/Model/UserVote.cs
@@ -1,0 +1,8 @@
+﻿namespace SongPlayHistory.Model
+{
+    internal class UserVote
+    {
+        public string key = null;
+        public string voteType = null;
+    }
+}
